fix: 불필요한 중계 리스너 제거
