+++ src/main/kotlin/kr/hhplus/be/server/domain/reservation/event/listener/ReservationCancelledEventListener.kt
@@ -1,7 +1,9 @@
-package kr.hhplus.be.server.domain.platform.event.listener
+package kr.hhplus.be.server.domain.reservation.event.listener
 
 import kr.hhplus.be.server.domain.reservation.event.ReservationCancelledEvent
 import kr.hhplus.be.server.global.client.ConcertDataPlatformClient
+import kr.hhplus.be.server.global.event.EventErrorHandler
+import kr.hhplus.be.server.global.event.EventErrorHandling
 import mu.KotlinLogging
 import org.springframework.context.event.EventListener
 import org.springframework.scheduling.annotation.Async
@@ -9,16 +11,17 @@ import org.springframework.stereotype.Component
 
 @Component
 class ReservationCancelledEventListener(
-    private val concertDataPlatformClient: ConcertDataPlatformClient
+    private val concertDataPlatformClient: ConcertDataPlatformClient,
+    private val errorHandler: EventErrorHandler
 ) {
     
     private val logger = KotlinLogging.logger {}
     
+    @EventErrorHandling(sendToDLQ = false, critical = false)
     @Async
     @EventListener
     fun handle(event: ReservationCancelledEvent) {
-        try {
-            // 데이터 플랫폼에 예약 취소 정보 전송 (비동기)
+        errorHandler.handleEventSafely(event, "ReservationCancelledEvent") {
             concertDataPlatformClient.sendReservationData(
                 reservationId = event.reservationId,
                 userId = event.userId,
@@ -26,11 +29,7 @@ class ReservationCancelledEventListener(
                 seatId = event.seatId,
                 operationType = "RESERVATION_CANCELLED"
             )
-            
-            logger.info { "데이터 플랫폼 예약 취소 정보 전송 완료 - reservationId: ${event.reservationId}" }
-            
-        } catch (e: Exception) {
-            logger.error(e) { "데이터 플랫폼 예약 취소 정보 전송 실패 - reservationId: ${event.reservationId}" }
+            logger.info { "데이터 플랫폼 예약 취소 정보 전송 완료: ${event.reservationId}" }
         }
     }
 }
